MultipleFileInterface
